+++ app/routes/content/index.js
@@ -150,7 +150,7 @@ export default Route.extend(Analytics, ResetScrollMixin, SetupSubmitControllerMi
         const dateCreated = new Date(preprint.get('dateCreated') || null);
         const dateModified = new Date(preprint.get('dateModified') || dateCreated);
         if (!preprint.get('datePublished')) { preprint.set('datePublished', dateCreated); }
-        const providerName = provider.get('name');
+        const providerName = provider.get('name') === 'Open Science Framework' ? 'OSF Preprints' : provider.get('name');
         const canonicalUrl = preprint.get('links.html');
         const contributors = this.get('contributors');
         const downloadUrl = this.get('downloadUrl');