+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Unified solution for preprint words
 - use of ember-osf `scheduled-banner` component
 - Selected provider description in the provider carousel
+- use of "OSF Preprints" as provider name for OSF preprints instead of "OSF" or "Open Science Framework"
 
 ### Changed
 - Styling and format of the branded navbar to match current styling in ember-osf